+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: (DOT) Add Subscan API key support
 - changed: (TON) Replace Orbs servers with dRPC
 
 ## 4.78.0 (2026-04-06)